[ci skip] MainSplitter: rename file extention to .h
